IPA protocol is deleted
Code compiles, but likely I removed more than required
