+++ ipa-core/src/protocol/mod.rs
@@ -1,14 +1,10 @@
-pub mod aggregation;
-pub mod attribution;
 pub mod basics;
 pub mod boolean;
 pub mod context;
 pub mod dp;
-pub mod ipa;
 pub mod ipa_prf;
 pub mod modulus_conversion;
 pub mod prss;
-pub mod sort;
 pub mod step;
 
 use std::{
