+++ ipa-core/src/protocol/step/compact.rs
@@ -34,16 +34,12 @@ impl Debug for Compact {
 }
 
 const ROOT_STATE: u16 = 0;
-const QUERY_TYPE_SEMIHONEST_STATE: u16 = 65533;
-const QUERY_TYPE_MALICIOUS_STATE: u16 = 65532;
-const PRSS_EXCHANGE_STATE: u16 = 65531;
-const QUERY_TYPE_OPRF_STATE: u16 = 65530;
+const QUERY_TYPE_OPRF_STATE: u16 = 65533;
+const PRSS_EXCHANGE_STATE: u16 = 65532;
 
 impl StepNarrow<QueryType> for Compact {
     fn narrow(&self, step: &QueryType) -> Self {
         match step {
-            QueryType::SemiHonestIpa(_) => Self(QUERY_TYPE_SEMIHONEST_STATE),
-            QueryType::MaliciousIpa(_) => Self(QUERY_TYPE_MALICIOUS_STATE),
             QueryType::OprfIpa(_) => Self(QUERY_TYPE_OPRF_STATE),
             _ => panic!("cannot narrow from the invalid step {}", step.as_ref()),
         }
@@ -60,8 +56,6 @@ impl StepNarrow<PrssExchangeStep> for Compact {
 fn static_reverse_state_map(state: u16) -> &'static str {
     match state {
         ROOT_STATE => "run-0",
-        QUERY_TYPE_SEMIHONEST_STATE => QueryType::SEMIHONEST_IPA_STR,
-        QUERY_TYPE_MALICIOUS_STATE => QueryType::MALICIOUS_IPA_STR,
         QUERY_TYPE_OPRF_STATE => QueryType::OPRF_IPA_STR,
         PRSS_EXCHANGE_STATE => PrssExchangeStep.as_ref(),
         _ => panic!("cannot as_ref() from the invalid state {state}"),
@@ -71,10 +65,6 @@ fn static_reverse_state_map(state: u16) -> &'static str {
 fn static_deserialize_state_map(s: &str) -> u16 {
     if s == "run-0" {
         return ROOT_STATE;
-    } else if s == QueryType::SEMIHONEST_IPA_STR {
-        return QUERY_TYPE_SEMIHONEST_STATE;
-    } else if s == QueryType::MALICIOUS_IPA_STR {
-        return QUERY_TYPE_MALICIOUS_STATE;
     } else if s == QueryType::OPRF_IPA_STR {
         return QUERY_TYPE_OPRF_STATE;
     } else if s == PrssExchangeStep.as_ref() {
