+++ ipa-core/src/query/processor.rs
@@ -534,10 +534,11 @@ mod tests {
             ff::{Field, Fp31},
             helpers::query::IpaQueryConfig,
             ipa_test_input,
-            protocol::{ipa::IPAInputRow, BreakdownKey, MatchKey},
+            protocol::{BreakdownKey, MatchKey},
             secret_sharing::replicated::semi_honest,
             test_fixture::{input::GenericReportTestInput, Reconstruct, TestApp},
         };
+        use crate::report::OprfReport;
 
         #[tokio::test]
         async fn complete_query_test_multiply() -> Result<(), BoxError> {
@@ -602,34 +603,35 @@ mod tests {
         }
 
         async fn ipa_query(app: &TestApp) -> Result<(), BoxError> {
-            let records: Vec<GenericReportTestInput<Fp31, MatchKey, BreakdownKey>> = ipa_test_input!(
-                [
-                    { timestamp: 0, match_key: 12345, is_trigger_report: 0, breakdown_key: 1, trigger_value: 0 },
-                    { timestamp: 0, match_key: 12345, is_trigger_report: 0, breakdown_key: 2, trigger_value: 0 },
-                    { timestamp: 0, match_key: 68362, is_trigger_report: 0, breakdown_key: 1, trigger_value: 0 },
-                    { timestamp: 0, match_key: 12345, is_trigger_report: 1, breakdown_key: 0, trigger_value: 5 },
-                    { timestamp: 0, match_key: 68362, is_trigger_report: 1, breakdown_key: 0, trigger_value: 2 },
-                ];
-                (Fp31, MatchKey, BreakdownKey)
-            );
-            let record_count = records.len();
-
-            let _results = app
-                .execute_query::<_, Vec<IPAInputRow<_, _, _>>>(
-                    records.into_iter(),
-                    QueryConfig {
-                        size: record_count.try_into().unwrap(),
-                        field_type: FieldType::Fp31,
-                        query_type: QueryType::SemiHonestIpa(IpaQueryConfig {
-                            per_user_credit_cap: 3,
-                            max_breakdown_key: 3,
-                            attribution_window_seconds: None,
-                            num_multi_bits: 3,
-                            plaintext_match_keys: true,
-                        }),
-                    },
-                )
-                .await?;
+            todo!("fix the input row stuff buddy");
+            // let records: Vec<GenericReportTestInput<Fp31, MatchKey, BreakdownKey>> = ipa_test_input!(
+            //     [
+            //         { timestamp: 0, match_key: 12345, is_trigger_report: 0, breakdown_key: 1, trigger_value: 0 },
+            //         { timestamp: 0, match_key: 12345, is_trigger_report: 0, breakdown_key: 2, trigger_value: 0 },
+            //         { timestamp: 0, match_key: 68362, is_trigger_report: 0, breakdown_key: 1, trigger_value: 0 },
+            //         { timestamp: 0, match_key: 12345, is_trigger_report: 1, breakdown_key: 0, trigger_value: 5 },
+            //         { timestamp: 0, match_key: 68362, is_trigger_report: 1, breakdown_key: 0, trigger_value: 2 },
+            //     ];
+            //     (Fp31, MatchKey, BreakdownKey)
+            // );
+            // let record_count = records.len();
+            //
+            // let _results = app
+            //     .execute_query::<_, Vec<OprfReport<_, _, _>>>(
+            //         records.into_iter(),
+            //         QueryConfig {
+            //             size: record_count.try_into().unwrap(),
+            //             field_type: FieldType::Fp31,
+            //             query_type: QueryType::OprfIpa(IpaQueryConfig {
+            //                 per_user_credit_cap: 3,
+            //                 max_breakdown_key: 3,
+            //                 attribution_window_seconds: None,
+            //                 num_multi_bits: 3,
+            //                 plaintext_match_keys: true,
+            //             }),
+            //         },
+            //     )
+            //     .await?;
 
             Ok(())
         }
